finish CRUD using cloudinary

+++ src/services/adminService.js
@@ -15,7 +15,13 @@ let getAllAdminService = () => {
     return new Promise(async (resolve, reject) => {
         try {
             const allAdmin = await db.Users.findAll({
-                where: { roleId: ['R1', 'R2'], isApproved: true }
+                where: { roleId: ['R1', 'R2'], isApproved: true },
+                order: [
+                    ['roleId', 'ASC'],
+                ],
+                include: [
+                    { model: db.Allcodes, as: 'roleData', attributes: ['valueEn', 'valueVn'] }
+                ],
             })
             if (allAdmin.length <= 0) {
                 resolve({
@@ -41,7 +47,7 @@ let getAllAdminNotApprovedService = () => {
     return new Promise(async (resolve, reject) => {
         try {
             const allAdmin = await db.Users.findAll({
-                where: { roleId: 'R1', isApproved: false }
+                where: { roleId: 'R2', isApproved: false }
             })
             if (allAdmin.length <= 0) {
                 resolve({
@@ -199,18 +205,19 @@ let createNewProductSevice = (body) => {
                     errMessage: 'Missing required parameter'
                 })
             }
-            const cloudinaryUpload = await cloudinary.uploader.upload(body.image, { folder: 'uploads' });
+            const cloudinaryUpload = await cloudinary.uploader.upload(body.image, { folder: 'uploads_product' });
             const imageUrl = cloudinaryUpload.secure_url;
-            // Lưu đường dẫn vào cơ sở dữ liệu bằng Sequelize
+            const cloudinaryId = cloudinaryUpload.public_id;
 
+            // Lưu đường dẫn va cloudID vào cơ sở dữ liệu bằng Sequelize
             await db.Products.create({
                 name: body.name,
                 originalPrice: body.originalPrice,
                 category: body.category,
-                // image: body.image,
                 image: imageUrl,
                 description: body.description,
-                quantitySold: 0
+                quantitySold: 0,
+                cloudId: cloudinaryId
             })
             resolve({
                 errCode: 0,
@@ -225,13 +232,31 @@ let createNewProductSevice = (body) => {
 let deleteProductService = (id) => {
     return new Promise(async (resolve, reject) => {
         try {
-            await db.Products.destroy({
+            const productDeleted = await db.Products.findOne({
                 where: { id: id }
             });
-            resolve({
-                errCode: 0,
-                errMessage: 'Delete product success'
-            })
+            if (productDeleted) {
+                const publicId = productDeleted.cloudId;
+                // // Xóa ảnh từ Cloudinary bằng public_id
+                const res = await cloudinary.uploader.destroy(publicId);
+                if (res.result === 'ok') {
+                    await productDeleted.destroy();
+                    resolve({
+                        errCode: 0,
+                        errMessage: 'Delete product success',
+                    })
+                } else {
+                    resolve({
+                        errCode: 1,
+                        errMessage: 'Delete product fail',
+                    })
+                }
+            } else {
+                resolve({
+                    errCode: 2,
+                    errMessage: 'Dont find product in database'
+                })
+            }
         } catch (e) {
             reject(e)
         }
@@ -248,19 +273,33 @@ let updateProductDataService = (body) => {
                 productUpdate.name = body.name
                 productUpdate.description = body.description
                 productUpdate.category = body.category
-                // productUpdate.size = body.size
-                productUpdate.image = body.image
                 productUpdate.originalPrice = body.originalPrice
-                await productUpdate.save();
-                resolve({
-                    errCode: 0,
-                    errMessage: "Update data success"
-                })
+                const publicId = productUpdate.cloudId;
+                // Xóa image cũ
+                const res = await cloudinary.uploader.destroy(publicId);
+                if (res.result === 'ok') {
+                    // Cập nhật image mới
+                    const cloudinaryUpload = await cloudinary.uploader.upload(body.image, { folder: 'uploads_product' });
+                    const imageUrl = cloudinaryUpload.secure_url;
+                    const cloudinaryId = cloudinaryUpload.public_id;
+                    productUpdate.image = imageUrl
+                    productUpdate.cloudId = cloudinaryId
+                    await productUpdate.save();
+                    resolve({
+                        errCode: 0,
+                        errMessage: "Update data success"
+                    })
+                } else {
+                    resolve({
+                        errCode: 2,
+                        errMessage: 'Update product fail',
+                    })
+                }
 
             } else {
                 resolve({
                     errCode: 1,
-                    errMessage: 'User not found'
+                    errMessage: 'Product not found'
                 })
             }
 
@@ -286,12 +325,14 @@ let createNewStoreService = (body) => {
                 const newArray = []
                 const imageDatas = [...body.image] // Mảng dữ liệu Base64 của các ảnh
                 const promises = imageDatas.map(async (imageData) => {
-                    const cloudinaryUpload = await cloudinary.uploader.upload(imageData.base64Image, { folder: 'uploads' });
+                    const cloudinaryUpload = await cloudinary.uploader.upload(imageData.base64Image, { folder: 'uploads_store' });
                     const imageUrl = cloudinaryUpload.secure_url;
+                    const cloudinaryId = cloudinaryUpload.public_id;
                     imageData = { ...imageData }
                     newArray.push({
                         storeId: store.id,
-                        image: imageUrl
+                        image: imageUrl,
+                        cloudId: cloudinaryId
                     })
                 })
                 await Promise.all(promises);  // Wait process 
@@ -318,13 +359,36 @@ let deleteStoreService = (id) => {
             await db.Stores.destroy({
                 where: { id: id }
             });
-            await db.ImageStore.destroy({
+            const storeDeleted = await db.ImageStore.findAll({
                 where: { storeId: id }
             })
-            resolve({
-                errCode: 0,
-                errMessage: 'Delete product success'
-            })
+            if (storeDeleted.length > 0) {
+                // Xóa nhiều ảnh từ Cloudinary bằng danh sách public_id
+                const deleteResults = await Promise.all(
+                    storeDeleted.map(async store => {
+                        public_id = store.cloudId
+                        const result = await cloudinary.uploader.destroy(public_id);
+                        return { public_id, result };
+                    })
+                );
+                const successfullyDeleted = deleteResults.filter(result => result.result === 'ok');
+                const failedToDelete = deleteResults.filter(result => result.result !== 'ok');
+
+                await db.ImageStore.destroy({
+                    where: { storeId: id }
+                })
+                resolve({
+                    errCode: 0,
+                    errMessage: 'Delete product success',
+                    successfullyDeleted,
+                    failedToDelete
+                })
+            } else {
+                resolve({
+                    errCode: 1,
+                    errMessage: 'Delete product fail',
+                })
+            }
         } catch (e) {
             reject(e)
         }
@@ -334,34 +398,68 @@ let deleteStoreService = (id) => {
 let updateStoreDataService = (body) => {
     return new Promise(async (resolve, reject) => {
         try {
-            await db.ImageStore.destroy({
+            const storeDeleted = await db.ImageStore.findAll({
                 where: { storeId: body.id }
             })
-            await uploadMultiImageService(body.image)
-            const storeUpdate = await db.Stores.findOne({
-                where: { id: body.id }
-            })
-            if (storeUpdate) {
-                storeUpdate.nameStore = body.nameStore
-                storeUpdate.address = body.address
-                storeUpdate.cityId = body.cityId
-                storeUpdate.description = body.description
-                storeUpdate.shortDescription = body.shortDescription
-                storeUpdate.mapLink = body.mapLink
-                storeUpdate.mapHTML = body.mapHTML
-                await storeUpdate.save();
-                resolve({
-                    errCode: 0,
-                    errMessage: "Update data success"
+            if (storeDeleted.length > 0) {
+                // Xóa nhiều ảnh từ Cloudinary bằng danh sách public_id
+                const deleteResults = await Promise.all(
+                    storeDeleted.map(async store => {
+                        public_id = store.cloudId
+                        const result = await cloudinary.uploader.destroy(public_id);
+                        return { public_id, result };
+                    })
+                );
+                const successfullyDeleted = deleteResults.filter(result => result.result === 'ok');
+                const failedToDelete = deleteResults.filter(result => result.result !== 'ok');
+
+                await db.ImageStore.destroy({
+                    where: { storeId: body.id }
+                })
+                const storeUpdate = await db.Stores.findOne({
+                    where: { id: body.id }
                 })
 
+                // Update các dữ liệu khác
+                if (storeUpdate) {
+                    storeUpdate.nameStore = body.nameStore
+                    storeUpdate.address = body.address
+                    storeUpdate.cityId = body.cityId
+                    storeUpdate.description = body.description
+                    storeUpdate.shortDescription = body.shortDescription
+                    storeUpdate.mapLink = body.mapLink
+                    storeUpdate.mapHTML = body.mapHTML
+                    await storeUpdate.save();
+
+                    // Đẩy nhiều ảnh mới lên Cloud
+                    const newArray = []
+                    const imageDatas = [...body.image] // Mảng dữ liệu Base64 của các ảnh
+                    const promises = imageDatas.map(async (imageData) => {
+                        const cloudinaryUpload = await cloudinary.uploader.upload(imageData.base64Image, { folder: 'uploads_store' });
+                        const imageUrl = cloudinaryUpload.secure_url;
+                        const cloudinaryId = cloudinaryUpload.public_id;
+                        imageData = { ...imageData }
+                        newArray.push({
+                            storeId: storeUpdate.id,
+                            image: imageUrl,
+                            cloudId: cloudinaryId
+                        })
+                    })
+                    await Promise.all(promises);  // Wait process 
+                    await db.ImageStore.bulkCreate(newArray)
+                    resolve({
+                        errCode: 0,
+                        errMessage: 'Update store success',
+                        successfullyDeleted,
+                        failedToDelete
+                    })
+                }
             } else {
                 resolve({
                     errCode: 1,
-                    errMessage: 'User not found'
+                    errMessage: 'Store not found'
                 })
             }
-
         } catch (e) {
             reject(e);
         }
@@ -404,7 +502,7 @@ let createNewManagerService = (body) => {
                     address: body.address,
                     roleId: 'R1', // role manager
                     phone: body.phone,
-                    isApproved: false
+                    isApproved: true
                 })
                 resolve({
                     errCode: 0,
